Added and improved tracing of a few instructions in MethodWalker
'ldnull', 'ldftn', 'newarr' and 'stelem.ref' were added.

'newobj' now pops arguments passed to the constructor and notifies the
OnCall handler of the constructor call.

'stfld' now pops the value to be stored: even though it is not passed
to the OnStore handler, leaving it on the stack can break the
interpretation of later instructions.