Add option to force extraction of enum values for specific type names
proto-extractor uses a series of heuristics to identify when a type
originates from Protobuf IDL, and enums don't inherently match any of those
heuristics. For most enums, that's not a problem, because there's a second
pass through all the types to identify any referenced types - if a Protobuf
message references a non-primitive type, that type must also be a Protobuf
type. In some cases, however, an enum type will not be referenced by any
other Protobuf types in an assembly.

The new command line argument `includeEnums` allows these types to be
specified explicitly so that they can be extracted alongside other
Protobuf-derived classes and enum types.